Terf wiki/requirements.txt
transformers>=4.40.0
torch>=2.0.0
sentence-transformers>=2.2.0
faiss-cpu>=1.7.4
requests>=2.28.0
discord.py>=2.3.0
huggingface_hub>=0.20.0
